added a liveness check; split out request specs into separate files
