+++ lib/routes/monitoring.rb
@@ -1,0 +1,3 @@
+get "/-/live" do
+  200
+end
